move charging test to tests mod
